Helper

Description

Contains small functions that are used in multiple places.

Documentation

bitsList :: [Int]

expandBits :: Word32 -> Word64

Returns a single 64-bit number where all even-indexed bits are set to the bits of a.

So the MSB is 0, the next bit is the MSB of a, the next is 0, the next is next-highest bit of a etc.

Kudos to https://lemire.me/blog/2018/01/08/how-fast-can-you-bit-interleave-32-bit-integers/ for explaining this bit-twiddling technique in detail.

interleaveBits :: Word32 -> Word32 -> Word64

Returns the morton-code interleaving

where from MSB to LSB we have one bit of y and then of x (and then the next-highest bit of y and then the next-highest bit of x etc).

shrinkBits :: Word64 -> Word32

deinterleaveBits :: Word64 -> (Word32, Word32)