Lib

Description

Main entry-point of the `domain model' part of the application. This module re-exports some of the nested module's types so that we can e.g. write `Lib.Camera === Lib.Camera.Camera`

Documentation

naivePointCloudToPicture :: Acc (Scalar Camera) -> Acc (Scalar Int) -> Acc (Scalar Int) -> Acc (Vector Point) -> Acc RasterPicture

type Camera = Transformation

A camera is just another transformation

type RasterPicture = Matrix Word32

type Transformation = M33 Float

An affine transformation is manipulated in this program as its augmented matrix.

type IFS = Vector (Transformation, Probability)

An Iterated Function System is described by a collection of transformations with, for the Chaos Game, associated probabilities