Lib.Camera

Description

transformations to (groups of) points to turn them from world-space to screen-space.

The camera (a transformation matrix) itself is meant to be manipulated on the CPU. The resulting transformation is then given to the GPU to map over all (visible) points.

Documentation

type Camera = Transformation

A camera is just another transformation

scale :: Float -> Camera -> Camera

Scales the camera in equal proportions using the given scale float.

Scales to the middle of the screen.

translate :: Float -> Float -> Camera -> Camera

Translates the camera position using the given horizontal and vertical offsets.

fromSixtuple :: (Float, Float, Float, Float, Float, Float) -> Camera

cameraTransform :: Camera -> Point -> Point

Transforms a point from world space to screen space.

cameraTransformGPU :: Exp Camera -> Exp Point -> Exp Point

Transforms a point from world space to screen space. (Runs on the GPU!)

inverse :: Camera -> Camera

Inverts the transformation the camera makes. useful for e.g. checking where the picure bounds (screen space) end up in world space.

withInitial :: Camera -> Camera -> Camera

identity :: Transformation