Lib.ChaosGame

Description

Parallel 'Chaos Game' implementation. The Chaos Game is a probabilistic algorithm to draw an Iterated Function System. Curiously, given enough points, the contraction of the IFS will ensure that the result will converge to the same deterministic picture.

This implementation of the Chaos Game takes a description of an IFS as input, performs all random number generation and point-transforming on the GPU itself, and finally returns a single vector of points as output.

This vector might be kept on the GPU for further manipulation (e.g. drawing pictures from them) afterwards.

Note that many internal functions of this module have also been exported, to make testing the various components easier.

Documentation

chaosGame

Arguments

:: Acc IFS
A description of the IFS' transformations + probabilities

-> Int
the amount of points to compute on each GPU thread.

-> Int
the number of threads to spin up.

-> Acc (Matrix Word64)
A random matrix

-> Acc (Vector Point)
A representation of the resulting point cloud.

Runs the chaos game

Note that this implementation runs on a two-dimensional matrix: each row is handled by a different thread each field in each row is the next iteration of the chaos game.

Essentially this means that for a high number of n_points_per_thread we'll have very high-quality points. whereas for a high paralellism we'll be able to run the chaosGame faster.

So depending on the particular IFS and GPU, you'll want to tune those values differently.

fillChaosGameMatrix :: Acc IFS -> Acc (Matrix Point) -> Acc (Matrix Point)

Given a matrix with random points we will perform the 'chaos game', separately on each row. So all points in column `0` will be initial points, in column `1` will have one transformation applied, in column `2` two transformations, etc.

floatPairToWord64 :: Exp Point -> Exp Word64

A fast way to turn a (Float, Float) into a Word64.

The hope is that LLVM or the PTX compilers might optimize it to a no-op, if the pair is adjacent in memory. Do not confuse this with pointToMorton.

word64ToFloatPair :: Exp Word64 -> Exp Point

Inverse of floatPairToWord64