Lib.Common

Description

Contains common types, type aliases and functions that are used throughout the application

Documentation

type Point = (Float, Float)

Since we are working in 2D, a point contaisn two coordinates.

type HomogeneousPoint = V3 Float

A point that we can easily transform using our Transformation type.

pointToHomogeneous :: Point -> HomogeneousPoint

Turns a 2D (x, y) point into a 3D vector ('homogeneous notation') where the z component is 1

pointToHomogeneousGPU :: Exp Point -> Exp HomogeneousPoint

Identical to pointToHomogeneous but runs on the GPU

homogeneousToPoint :: HomogeneousPoint -> Point

Inverse of pointToHomogeneous

Note that s is not always `1`; it is altered by scaling transformations.

homogeneousToPointGPU :: Exp HomogeneousPoint -> Exp Point

Identical to homogeneousToPoint but runs on the GPU

mapPointAsHomogeneous :: (HomogeneousPoint -> HomogeneousPoint) -> Point -> Point

Runs a function working on a point in homogeneous form on a normal 2D point. by wrapping it in a conversion to and from homogeneous notation

mapPointAsHomogeneousGPU :: (Exp HomogeneousPoint -> Exp HomogeneousPoint) -> Exp Point -> Exp Point

Identical to mapPointAsHomogeneous but runs on the GPU