Lib.Guide

Description

Module to render guideś. These are four-point polygons which indicate (rough estimates of) regions in which all points have had a particular transformation applied to them.

Documentation

drawGuides

Arguments

:: Camera
The current camera matrix

-> Camera
The initial camera matrix

-> (Word, Word)
The screen dimensions

-> [Transformation]
List of transformations of this IFS

-> Picture

allCombinations :: [a] -> [[a]]

Returns all combinations of the elements in a list This goes on forever. Shorter combinations are first

## Examples:

>>> take 4 $ allCombinations [1, 2, 3]
[[],[1],[2],[3]]
>>> take 13 $ allCombinations [1, 2, 3]
[[],[1],[2],[3],[1,1],[1,2],[1,3],[2,1],[2,2],[2,3],[3,1],[3,2],[3,3]]

combinationsUpToDepth :: Int -> [a] -> [[a]]

Returns all combinations of the elements in a list that are at most depth elements long Shorter combinations are returned first.

## Examples:

>>> combinationsUpToDepth 0 [1, 2, 3]
[[]]
>>> combinationsUpToDepth 1 [1, 2, 3]
[[],[1],[2],[3]]
>>> combinationsUpToDepth 2 [1, 2, 3]
[[],[1],[2],[3],[1,1],[1,2],[1,3],[2,1],[2,2],[2,3],[3,1],[3,2],[3,3]]

isCameraInsideTransformation :: Camera -> Camera -> Transformation -> Bool

Checks whether the current camera is inside a transformation

Here inside means that the screen transformed by the inverse camera matrix, is completely contained in the screen transformed by the transformation.

initial_camera is used to move between 'normalized screen coordinates' (which camera uses) and 'world coordinates' (which transformation uses).