Lib.Picture

Description

Transforming a point cloud to a rasterized picture, using a camera transformation.

Documentation

type RasterPicture = Matrix Word32

naivePointCloudToPicture :: Exp Camera -> Exp Int -> Exp Int -> Acc (Vector Point) -> Acc RasterPicture

A very simple way to create a picture from a point cloud. No optimizations: we iterate over all points and only find out very late whether they'll end up on the screen.