Lib.Random

Description

Generate random data in parallel on the GPU, based on a single seed.

We use the xorShift method to generate data. This is not a very high-quality generator, but it should be good enough for our purposes.

Documentation

randomVector :: Int -> Word64 -> Acc (Vector Word64)

Fills a vector of length length with random numbers using seed as starting seed.

randomMatrix :: Int -> Int -> Word64 -> Acc (Matrix Word64)

A cute implementation of a random matrix.

However, since we do not fully split the RNG, some entries to the top-left of the matrix might be close to each-other.

If better RNG is needed, we might be able to use something like the jump mentioned on http://prng.di.unimi.it/

xorShift64 :: Exp Word64 -> Exp Word64

Simple XorShift Random Number Generator Returns a new number based on the given one. Note that XorShift is an _ok_ RNG but not a very high-quality one: Specifically, it will fail certain statistical tests. Also, it will never return `0` (if given any number except `0`).

extractRandomUnitIntervalDouble :: Exp Word64 -> Exp Double

Returns a double in the half-open range [0, 1) based on the random unsigned integer that was generated.

This uses the notes at http://prng.di.unimi.it/. Specifically: A double has a significant binary digits 53. We extract thus the top 53 bits of the number (which are better quality than the lower 11) We multiply this by `0x1.0p-53` (that is: 1.0 * 2^-53) to obtain a number in the half-open interval [0, 1)