Lib.Sort

Description

This module deals with the sorting of data in parallel on the GPU.

The main entrypoint is sortPoints. The other exposed functions are building blocks of that one. They are exposed to be easily testable themselves.

Documentation

sortPoints :: Acc (Vector (Float, Float)) -> Acc (Vector (Float, Float))

Sorts an array of points by the Z-order curve (AKA morton code)

This is done by transforming the pair of 32-bit floats to 32-bit unsigned int, and combining them to get one unsigned 64-bit int. After sorting, this transformation is reversed.

radixSort :: Acc (Vector Word64) -> Acc (Vector Word64)

A full parallel Radix Sort

## Implementation notes:

This implementation uses 64 unrolled implementations of radixSortBit, one for each bit in a 64-bit number.

The nice thing about using Accelerate is that we can manipulate the AST creation using techniques like this one where we can create an unrolled version of something that would be horrible to write by hand.

## Examples

>>> import qualified Data.Array.Accelerate.LLVM.Native as CPU
>>> arr = use $ fromList (Z :. 10) ([10, 9..]) :: Acc (Vector Word64)
>>> CPU.run $ Lib.Sort.radixSort arr
Vector (Z :. 10) [1,2,3,4,5,6,7,8,9,10]

radixSortBit :: Int -> Acc (Vector Word64) -> Acc (Vector Word64)

One step of parallel radix sort, for a single bit.