Lib.Transformation

Description

Hides the implementation details of transformations. To the outside world they are opaque, and only the operations we require are exported from this module.

Secretly they are 3x3 matrices.

Documentation

type Transformation = M33 Float

An affine transformation is manipulated in this program as its augmented matrix.

type Probability = Float

Type that indicates that we are expecting a float in the [0..1) range.

type IFS = Vector (Transformation, Probability)

An Iterated Function System is described by a collection of transformations with, for the Chaos Game, associated probabilities

fromSixtuplePair :: ((Float, Float, Float, Float, Float, Float), Probability) -> (Transformation, Probability)

fromSixtuplePairGPU :: Exp ((Float, Float, Float, Float, Float, Float), Probability) -> Exp (Transformation, Probability)

fromSixtuple :: (Float, Float, Float, Float, Float, Float) -> Transformation

fromSixtupleGPU :: Exp (Float, Float, Float, Float, Float, Float) -> Exp Transformation

identity :: Transformation

isInvertible :: Transformation -> Bool

transform :: Transformation -> Point -> Point

transformGPU :: Exp Transformation -> Exp Point -> Exp Point

Transforms a point from world space to screen space. (Runs on the GPU!)

invert :: Transformation -> Maybe Transformation

combine :: [Transformation] -> Transformation